Adding a manually built OPML file
I wanted to get the feeds for all the sites listed.

Unfortunately, many didn't have feeds, specially those using gatsby. Further, some were using a custom framework or just weren't exposing feeds. I was able to garner some more with guess-work, but many of them are comments because I just couldn't find feeds.

It's not necessary for people to fill out this OPML file when they send the PR, but it sure would be nice if they did.