Add Leo Van's Blog
This my personal blog in both Chinese and English. Now, homepage, about page and resume page is available in English, more English posts are on the way. Most of posts are machine learning related, and some are my opinions on books, movies and life. Currently, I am working on intelligent agriculture, thus it will get more posts in this subject.